// src/api/model/formField.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Form API
 * API for creating and managing forms
 * OpenAPI spec version: 1.0.0
 */
import type { FormFieldType } from "./formFieldType";
import type { FormFieldFieldType } from "./formFieldFieldType";

export interface FormField {
  _id?: string;
  /** Unique identifier for the field */
  id?: string;
  type?: FormFieldType;
  /** Type of the form field */
  fieldType?: FormFieldFieldType;
  title?: string;
  options?: string[];
  order?: number;
  required?: boolean;
}