// src/api/model/formFieldType.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Form API
 * API for creating and managing forms
 * OpenAPI spec version: 1.0.0
 */

export type FormFieldType = (typeof FormFieldType)[keyof typeof FormFieldType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FormFieldType = {
  "short-text": "short-text",
  "long-text": "long-text",
  "multiple-choice": "multiple-choice",
  dropdown: "dropdown",
  checkbox: "checkbox",
} as const;