// src/api/model/formRequest.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Form API
 * API for creating and managing forms
 * OpenAPI spec version: 1.0.0
 */
import type { FormField } from "./formField";
import type { FormSettings } from "./formSettings";
import type { FormRequestStatus } from "./formRequestStatus";

export interface FormRequest {
  /** Title of the form */
  title?: string;
  /** Description of the form */
  description?: string;
  /** List of form fields */
  fields?: FormField[];
  settings?: FormSettings;
  /** Status of the form */
  status?: FormRequestStatus;
}